Revert "extracted MemoryPollingThread"
This reverts commit a88f7825121b240a5d359dd671aa81b3ddf1b6f2.
