+++ micrometer-core/src/main/java/io/micrometer/core/instrument/AbstractTimer.java
@@ -18,6 +18,7 @@ package io.micrometer.core.instrument;
 import java.lang.management.ManagementFactory;
 import java.lang.management.MemoryMXBean;
 import java.lang.management.MemoryNotificationInfo;
+import java.lang.management.MemoryPoolMXBean;
 import java.util.Map;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ConcurrentHashMap;
@@ -120,6 +121,7 @@ public abstract class AbstractTimer extends AbstractMeter implements Timer {
                     @Override
                     public void handleNotification(Notification notification, Object handback) {
                         if (notification.getType().equals(MemoryNotificationInfo.MEMORY_THRESHOLD_EXCEEDED)) {
+                            System.out.println("Fermata: " + notification.getType() + ", PAUSE ALL OBSERVABILITY THREADS");
                             new MemoryPollingThread(memThresholdDetector.getPollMemoryPoolThresholdMillis()).start();
                         }
                         System.out.println("Separate thread launched handleNotification() complete!");
@@ -147,6 +149,39 @@ public abstract class AbstractTimer extends AbstractMeter implements Timer {
                     }
                 }
             });
+        }
+    }
+
+    class MemoryPollingThread extends Thread {
+        private int pollMemoryThresholdMillis = 6000;
+        
+        public MemoryPollingThread(int pollMemoryThresholdMillis) {
+            this.pollMemoryThresholdMillis = pollMemoryThresholdMillis;
+        }
+
+        public void run() {
+            //poll indefinitely every (default) 6 seconds to determine if all memory pools are freed up
+            while (isMemoryPoolThresholdExceeded()) {
+                try {
+                    System.out.println("indefinitely looping every " + pollMemoryThresholdMillis + " millis!");
+                    Thread.sleep(pollMemoryThresholdMillis);
+                } catch (InterruptedException e) {
+                    e.printStackTrace();
+                }
+            }
+            System.out.println("Tremolo: exited polling loop, CONTINUE ALL OBSERVABILITY THREADS!");
+        }
+        
+        private boolean isMemoryPoolThresholdExceeded() {
+            for (MemoryPoolMXBean membean : ManagementFactory.getMemoryPoolMXBeans()) {
+                if (membean.isUsageThresholdSupported()) {
+                    System.out.println("name: " + membean.getName());
+                    if (membean.isUsageThresholdExceeded()) {
+                        return true;
+                    }
+                }
+            }
+            return false;
         }
     }
