+++ micrometer-core/src/test/java/io/micrometer/core/instrument/binder/kafka/KafkaClientMetricsIntegrationTest.java
@@ -15,13 +15,7 @@
  */
 package io.micrometer.core.instrument.binder.kafka;
 
-import static java.lang.System.out;
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.time.Duration;
-import java.util.Collections;
-import java.util.Properties;
-
+import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
 import org.apache.kafka.clients.consumer.Consumer;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
@@ -37,9 +31,14 @@ import org.testcontainers.containers.KafkaContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
-import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
+import java.time.Duration;
+import java.util.Collections;
+import java.util.Properties;
 
-@Testcontainers(disabledWithoutDocker = true)
+import static java.lang.System.out;
+import static org.assertj.core.api.Assertions.assertThat;
+
+@Testcontainers
 @Tag("docker")
 class KafkaClientMetricsIntegrationTest {
     @Container
